+++ src/deepl/glossary_info.cr
@@ -8,7 +8,7 @@ module DeepL
     property source_lang : String
     property target_lang : String
     property creation_time : Time
-    property entry_count : UInt32
+    property entry_count : Int64
 
     def initialize(@glossary_id, @name, @ready, @source_lang, @target_lang, @creation_time, @entry_count)
     end
